requirements.txt
fastapi
python-jose[cryptography]
passlib[bcrypt]
pydantic
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pytest
pytest-asyncio
httpx
pytest-cov
pytest-mock
passlib
python-jose
bcrypt
uploadfiles
pdfminer.six
python-docx


